Fix experimentation printing problem

+++ Code/multiprocess_lab.py
@@ -35,7 +35,8 @@ Operators = ['Darwinian phi rate', 'evaluated insertion']
 #          'evaluated insertion': {e_keys[i]: e_combination[i] for i in range(len(e_keys))}, 
 #          'genetic parameters': {g_keys[i]: g_combination[i] for i in range(len(g_keys))}} for D_combination in D_combinations for e_combination in e_combinations for g_combination in g_combinations]
 
-Grid = [{'Darwinian phi rate':{'penalization':'cuadratic', 'conservation proportion':0.4, 'length restriction':False }, #1
+Grid = [
+         {'Darwinian phi rate':{'penalization':'cuadratic', 'conservation proportion':0.4, 'length restriction':False }, #1
          'evaluated insertion':{'penalization':'cubic', 'criterion':'random'},
          'genetic parameters':{'population size':1500, 'crossover rate': 0.6, 'mutation rate':0.6}},
          
@@ -87,25 +88,25 @@ test_batch = env.sizes['l']
 
 
 if __name__ == '__main__':
-    for num, Configs in enumerate(Grid[::-1]): 
-            num = 9-num
-            with open(path + f'Experimentation/Second phase (l)/Exp {num}/readme.txt', 'w') as f:
-                readme = f'Experiment {num}'
-                readme += f'\nDarwinian phi rate: \t{Configs["Darwinian phi rate"]["penalization"]} - {Configs["Darwinian phi rate"]["length restriction"]}'
-                readme += f'\nevaluated insertion: \t{Configs["evaluated insertion"]["penalization"]} - {Configs["evaluated insertion"]["criterion"]}'
-                readme += f'\ngenetic configuration: \t{Configs["genetic parameters"]["population size"]} - {Configs["genetic parameters"]["crossover rate"]} - {Configs["genetic parameters"]["mutation rate"]}'
-                f.write(readme)
-                    
+    for num, Configs in enumerate(Grid): 
+            if num > 0 and < 9:
+                with open(path + f'Experimentation/Second phase (l)/Exp {num}/readme.txt', 'w') as f:
+                    readme = f'Experiment {num}'
+                    readme += f'\nDarwinian phi rate: \t{Configs["Darwinian phi rate"]["penalization"]} - {Configs["Darwinian phi rate"]["length restriction"]}'
+                    readme += f'\nevaluated insertion: \t{Configs["evaluated insertion"]["penalization"]} - {Configs["evaluated insertion"]["criterion"]}'
+                    readme += f'\ngenetic configuration: \t{Configs["genetic parameters"]["population size"]} - {Configs["genetic parameters"]["crossover rate"]} - {Configs["genetic parameters"]["mutation rate"]}'
+                    f.write(readme)
+                        
 
-            progress_percentage = round(round((num+1)/len(Grid),4)*100,2)
-            print(f'\n-------- Experiment {num} / {progress_percentage}% --------')
+                progress_percentage = round(round((num+1)/len(Grid),4)*100,2)
+                print(f'\n-------- Experiment {num} / {progress_percentage}% --------')
 
-            lab:Experiment = Experiment(path, Operators, Configs, False, True, num)
+                lab:Experiment = Experiment(path, Operators, Configs, False, True, num)
 
-            if computer == 'mac':   p = pool.Pool(processes = 8)
-            else: p = pool.Pool()
+                if computer == 'mac':   p = pool.Pool(processes = 8)
+                else: p = pool.Pool()
 
-            Results = p.map(lab.experimentation, test_batch)
-            print(f'Average gap: {sum(Results)/len(Results)}')
+                Results = p.map(lab.experimentation, test_batch)
+                print(f'Average gap: {sum(Results)/len(Results)}')
 
-            p.terminate()
+                p.terminate()
